line-bot/requirements.txt
# Google Functions Framework for Python
# https://github.com/GoogleCloudPlatform/functions-framework-python
functions-framework==3.*

# Cloud Storage client library
# https://cloud.google.com/python/docs/reference/storage/latest
google-cloud-storage

# LINE Message API SDK for Python
# https://github.com/line/line-bot-sdk-python
line-bot-sdk
